requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1